% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data_NouvAqui.R
\name{import_data_NouvAqui}
\alias{import_data_NouvAqui}
\title{Import COVID-19 data from regional report}
\source{
Source: ARS Nouvelle Aquitaine
}
\usage{
import_data_NouvAqui(file_xlsx)
}
\description{
Import COVID-19 data from regional report
}
