% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Launch Covid19FRNA Shiny App}
\usage{
run_app(...)
}
\arguments{
\item{...}{additional arguments to be passed to the \link[shiny]{runApp} function.}
}
\description{
Launch Covid19FRNA Shiny App
}
\examples{
if(interactive()){
Covid19FRNA::run_app()
}

}
