% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrap_pe_spf.R
\name{scrap_pe_spf}
\alias{scrap_pe_spf}
\title{Extract data from the daily "point epidemiologique COVID-19"}
\source{
Source: Sante Publique France
}
\usage{
scrap_pe_spf()
}
\description{
Extract data from the daily "point epidemiologique COVID-19"
}
